requirements.txt
accelerate==1.12.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
filelock==3.20.2
Flask==3.1.2
flask-cors==6.0.2
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
mysql-connector-python==9.5.0
neo4j==6.0.3
networkx==3.6.1
numpy==2.4.0
packaging==25.0
psutil==7.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
